+++ image_substitutors_test.go
@@ -1,3 +1,5 @@
+// This test is testing very internal logic that should not be exported away from this package. We'll
+// leave it in the main testcontainers package. Do not use for user facing examples.
 package testcontainers
 
 import (
