+++ logconsumer_internal_test.go
@@ -1,0 +1,380 @@
+// This test is testing very internal logic that should not be exported away from this package. We'll
+// leave it in the main testcontainers package. Do not use for user facing examples.
+package testcontainers
+
+import (
+	"bytes"
+	"context"
+	"errors"
+	"fmt"
+	"net/http"
+	"os"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/docker/docker/client"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/testcontainers/testcontainers-go/internal/config"
+	"github.com/testcontainers/testcontainers-go/wait"
+)
+
+const lastMessage = "DONE"
+
+type TestLogConsumer struct {
+	Msgs []string
+	Done chan bool
+
+	// Accepted provides a blocking way of ensuring the logs messages have been consumed.
+	// This allows for proper synchronization during Test_StartStop in particular.
+	// Please see func devNullAcceptorChan if you're not interested in this synchronization.
+	Accepted chan string
+}
+
+func (g *TestLogConsumer) Accept(l Log) {
+	s := string(l.Content)
+	if s == fmt.Sprintf("echo %s\n", lastMessage) {
+		g.Done <- true
+		return
+	}
+	g.Accepted <- s
+	g.Msgs = append(g.Msgs, s)
+}
+
+// devNullAcceptorChan returns string channel that essentially sends all strings to dev null
+func devNullAcceptorChan() chan string {
+	c := make(chan string)
+	go func(c <-chan string) {
+		for range c {
+			// do nothing, just pull off channel
+		}
+	}(c)
+	return c
+}
+
+func TestContainerLogWithErrClosed(t *testing.T) {
+	if os.Getenv("XDG_RUNTIME_DIR") != "" {
+		t.Skip("Skipping as flaky on GitHub Actions, Please see https://github.com/testcontainers/testcontainers-go/issues/1924")
+	}
+
+	t.Cleanup(func() {
+		config.Reset()
+	})
+
+	if providerType == ProviderPodman {
+		t.Skip("Docker-in-Docker does not work with rootless Podman")
+	}
+	// First spin up a docker-in-docker container, then spin up an inner container within that dind container
+	// Logs are being read from the inner container via the dind container's tcp port, which can be briefly
+	// closed to test behaviour in connection-closed situations.
+	ctx := context.Background()
+
+	dind, err := GenericContainer(ctx, GenericContainerRequest{
+		Started: true,
+		ContainerRequest: ContainerRequest{
+			Image:        "docker.io/docker:dind",
+			ExposedPorts: []string{"2375/tcp"},
+			Env:          map[string]string{"DOCKER_TLS_CERTDIR": ""},
+			WaitingFor:   wait.ForListeningPort("2375/tcp"),
+			Privileged:   true,
+		},
+	})
+
+	require.NoError(t, err)
+	terminateContainerOnEnd(t, ctx, dind)
+
+	var remoteDocker string
+
+	ctxEndpoint, cancel := context.WithTimeout(ctx, 5*time.Second)
+	defer cancel()
+
+	// todo: remove this temporary fix (test is flaky).
+	for {
+		remoteDocker, err = dind.Endpoint(ctxEndpoint, "2375/tcp")
+		if err == nil {
+			break
+		}
+		if errors.Is(err, context.DeadlineExceeded) {
+			break
+		}
+		time.Sleep(100 * time.Microsecond)
+		t.Log("retrying get endpoint")
+	}
+	if err != nil {
+		t.Fatal("get endpoint:", err)
+	}
+
+	opts := []client.Opt{client.WithHost(remoteDocker), client.WithAPIVersionNegotiation()}
+
+	client, err := NewDockerClientWithOpts(ctx, opts...)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer client.Close()
+
+	provider := &DockerProvider{
+		client: client,
+		config: ReadConfig(),
+		DockerProviderOptions: &DockerProviderOptions{
+			GenericProviderOptions: &GenericProviderOptions{
+				Logger: TestLogger(t),
+			},
+		},
+	}
+
+	consumer := TestLogConsumer{
+		Msgs:     []string{},
+		Done:     make(chan bool),
+		Accepted: devNullAcceptorChan(),
+	}
+
+	nginx, err := provider.CreateContainer(ctx, ContainerRequest{
+		Image:        "nginx",
+		ExposedPorts: []string{"80/tcp"},
+		LogConsumerCfg: &LogConsumerConfig{
+			Consumers: []LogConsumer{&consumer},
+		},
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := nginx.Start(ctx); err != nil {
+		t.Fatal(err)
+	}
+	terminateContainerOnEnd(t, ctx, nginx)
+
+	port, err := nginx.MappedPort(ctx, "80/tcp")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Gather the initial container logs
+	time.Sleep(time.Second * 1)
+	existingLogs := len(consumer.Msgs)
+
+	hitNginx := func() {
+		i, _, err := dind.Exec(ctx, []string{"wget", "--spider", "localhost:" + port.Port()})
+		if err != nil || i > 0 {
+			t.Fatalf("Can't make request to nginx container from dind container")
+		}
+	}
+
+	hitNginx()
+	time.Sleep(time.Second * 1)
+	if len(consumer.Msgs)-existingLogs != 1 {
+		t.Fatalf("logConsumer should have 1 new log message, instead has: %v", consumer.Msgs[existingLogs:])
+	}
+	existingLogs = len(consumer.Msgs)
+
+	iptableArgs := []string{
+		"INPUT", "-p", "tcp", "--dport", "2375",
+		"-j", "REJECT", "--reject-with", "tcp-reset",
+	}
+	// Simulate a transient closed connection to the docker daemon
+	i, _, err := dind.Exec(ctx, append([]string{"iptables", "-A"}, iptableArgs...))
+	if err != nil || i > 0 {
+		t.Fatalf("Failed to close connection to dind daemon: i(%d), err %v", i, err)
+	}
+	i, _, err = dind.Exec(ctx, append([]string{"iptables", "-D"}, iptableArgs...))
+	if err != nil || i > 0 {
+		t.Fatalf("Failed to re-open connection to dind daemon: i(%d), err %v", i, err)
+	}
+	time.Sleep(time.Second * 3)
+
+	hitNginx()
+	hitNginx()
+	time.Sleep(time.Second * 1)
+	if len(consumer.Msgs)-existingLogs != 2 {
+		t.Fatalf(
+			"LogConsumer should have 2 new log messages after detecting closed connection and"+
+				" re-requesting logs. Instead has:\n%s", consumer.Msgs[existingLogs:],
+		)
+	}
+}
+
+func Test_StartLogProductionStillStartsWithTooLowTimeout(t *testing.T) {
+	ctx := context.Background()
+
+	g := TestLogConsumer{
+		Msgs:     []string{},
+		Done:     make(chan bool),
+		Accepted: devNullAcceptorChan(),
+	}
+
+	req := ContainerRequest{
+		FromDockerfile: FromDockerfile{
+			Context:    "./testdata/",
+			Dockerfile: "echoserver.Dockerfile",
+		},
+		ExposedPorts: []string{"8080/tcp"},
+		WaitingFor:   wait.ForLog("ready"),
+		LogConsumerCfg: &LogConsumerConfig{
+			Opts:      []LogProductionOption{WithLogProductionTimeout(4 * time.Second)},
+			Consumers: []LogConsumer{&g},
+		},
+	}
+
+	gReq := GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	}
+
+	c, err := GenericContainer(ctx, gReq)
+	require.NoError(t, err)
+	terminateContainerOnEnd(t, ctx, c)
+}
+
+func Test_StartLogProductionStillStartsWithTooHighTimeout(t *testing.T) {
+	ctx := context.Background()
+
+	g := TestLogConsumer{
+		Msgs:     []string{},
+		Done:     make(chan bool),
+		Accepted: devNullAcceptorChan(),
+	}
+
+	req := ContainerRequest{
+		FromDockerfile: FromDockerfile{
+			Context:    "./testdata/",
+			Dockerfile: "echoserver.Dockerfile",
+		},
+		ExposedPorts: []string{"8080/tcp"},
+		WaitingFor:   wait.ForLog("ready"),
+		LogConsumerCfg: &LogConsumerConfig{
+			Opts:      []LogProductionOption{WithLogProductionTimeout(61 * time.Second)},
+			Consumers: []LogConsumer{&g},
+		},
+	}
+
+	gReq := GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	}
+
+	c, err := GenericContainer(ctx, gReq)
+	require.NoError(t, err)
+	require.NotNil(t, c)
+
+	// because the log production timeout is too high, the container should have already been terminated
+	// so no need to terminate it again with "terminateContainerOnEnd(t, ctx, c)"
+	dc := c.(*DockerContainer)
+	require.NoError(t, dc.stopLogProduction())
+
+	terminateContainerOnEnd(t, ctx, c)
+}
+
+func Test_MultiContainerLogConsumer_CancelledContext(t *testing.T) {
+	// Redirect stderr to a buffer
+	oldStderr := os.Stderr
+	r, w, _ := os.Pipe()
+	os.Stderr = w
+
+	// Context with cancellation functionality for simulating user interruption
+	ctx, cancel := context.WithCancel(context.Background())
+
+	first := TestLogConsumer{
+		Msgs:     []string{},
+		Done:     make(chan bool),
+		Accepted: devNullAcceptorChan(),
+	}
+
+	containerReq1 := ContainerRequest{
+		FromDockerfile: FromDockerfile{
+			Context:    "./testdata/",
+			Dockerfile: "echoserver.Dockerfile",
+		},
+		ExposedPorts: []string{"8080/tcp"},
+		WaitingFor:   wait.ForLog("ready"),
+		LogConsumerCfg: &LogConsumerConfig{
+			Consumers: []LogConsumer{&first},
+		},
+	}
+
+	genericReq1 := GenericContainerRequest{
+		ContainerRequest: containerReq1,
+		Started:          true,
+	}
+
+	c, err := GenericContainer(ctx, genericReq1)
+	require.NoError(t, err)
+
+	ep1, err := c.Endpoint(ctx, "http")
+	require.NoError(t, err)
+
+	_, err = http.Get(ep1 + "/stdout?echo=hello1")
+	require.NoError(t, err)
+
+	_, err = http.Get(ep1 + "/stdout?echo=there1")
+	require.NoError(t, err)
+
+	second := TestLogConsumer{
+		Msgs:     []string{},
+		Done:     make(chan bool),
+		Accepted: devNullAcceptorChan(),
+	}
+
+	containerReq2 := ContainerRequest{
+		FromDockerfile: FromDockerfile{
+			Context:    "./testdata/",
+			Dockerfile: "echoserver.Dockerfile",
+		},
+		ExposedPorts: []string{"8080/tcp"},
+		WaitingFor:   wait.ForLog("ready"),
+		LogConsumerCfg: &LogConsumerConfig{
+			Consumers: []LogConsumer{&second},
+		},
+	}
+
+	genericReq2 := GenericContainerRequest{
+		ContainerRequest: containerReq2,
+		Started:          true,
+	}
+
+	c2, err := GenericContainer(ctx, genericReq2)
+	require.NoError(t, err)
+
+	ep2, err := c2.Endpoint(ctx, "http")
+	require.NoError(t, err)
+
+	_, err = http.Get(ep2 + "/stdout?echo=hello2")
+	require.NoError(t, err)
+
+	_, err = http.Get(ep2 + "/stdout?echo=there2")
+	require.NoError(t, err)
+
+	// Handling the termination of the containers
+	defer func() {
+		shutdownCtx, shutdownCancel := context.WithTimeout(
+			context.Background(), 60*time.Second,
+		)
+		defer shutdownCancel()
+		_ = c.Terminate(shutdownCtx)
+		_ = c2.Terminate(shutdownCtx)
+	}()
+
+	// Deliberately calling context cancel
+	cancel()
+
+	// We check log size due to context cancellation causing
+	// varying message counts, leading to test failure.
+	assert.GreaterOrEqual(t, len(first.Msgs), 2)
+	assert.GreaterOrEqual(t, len(second.Msgs), 2)
+
+	// Restore stderr
+	w.Close()
+	os.Stderr = oldStderr
+
+	// Read the stderr output from the buffer
+	var buf bytes.Buffer
+	_, _ = buf.ReadFrom(r)
+
+	// Check the stderr message
+	actual := buf.String()
+
+	// The context cancel shouldn't cause the system to throw a
+	// logStoppedForOutOfSyncMessage, as it hangs the system with
+	// the multiple containers.
+	assert.False(t, strings.Contains(actual, logStoppedForOutOfSyncMessage))
+}
